micro_servicios/etl_ibm/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
functions-framework
pyarrow